+++ tests/unit/test_cell_id.py
@@ -1,10 +1,8 @@
-from nbformat.v4.nbbase import new_code_cell
+from nbformat.v4.nbbase import new_code_cell, new_markdown_cell, new_raw_cell
 
 
 def test_cell_id_is_not_random():
-    id1 = new_code_cell().id
-    id2 = new_code_cell().id
-
-    n1 = int(id1.split("-")[1])
-    n2 = int(id2.split("-")[1])
-    assert n2 == n1 + 1
+    assert new_code_cell().id == "cell-1"
+    assert new_code_cell().id == "cell-2"
+    assert new_markdown_cell().id == "cell-3"
+    assert new_raw_cell().id == "cell-4"