+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Jupytext ChangeLog
 ==================
 
+1.16.2-dev (2024-03-??)
+-------------------
+
+**Changed**
+- Temporary text notebooks for the `--pipe` or `--check` commands are now created in the notebook directory ([#1206](https://github.com/mwouts/jupytext/issues/1206))
+
+**Fixed**
+- Jupytext is now tested with `pandoc>=3.0`. Please note that switching to `pandoc>=3.0` will add cell ids to your `pandoc:md` notebooks ([#1006](https://github.com/mwouts/jupytext/issues/1006))
+
+
 1.16.1 (2024-01-13)
 -------------------